% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD_example.R
\docType{data}
\name{AD_example}
\alias{AD_example}
\title{Example data of atopic dermatitis patients}
\format{
A matrix with 31688 rows (genes) and 20 variables (patients):
}
\source{
DEFINE SOURCE! (MAARS, pub, etc.)
}
\usage{
AD_example
}
\description{
Standardized microarray gene expression data of human skin from atopic dermatitis patients.
}
\keyword{datasets}
