% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_mods_enrichr.R
\name{ann_mods_enrichr}
\alias{ann_mods_enrichr}
\title{Pearson co-expression network Annotate modules with enrichR package}
\usage{
ann_mods_enrichr(in_graph, databs, outnam)
}
\arguments{
\item{in_graph}{(Required) igraph object with module membership information}

\item{databs}{(Required) enrich database to query, e.g. "GO_Biological_Process_2021".}

\item{outnam}{(Required) output base filename of significant enrichment terms table}
}
\description{
Annotate gene co-expression network modules, write enrichment term table to working directory.
Specific to Comean::construct_conet output
}
\examples{
ann_mods_enrichr()
}
\keyword{co-expression}
\keyword{enrichr}
\keyword{igraph}
\keyword{network}
