% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Removes all vertex attributes except for default or user-specified attributes.}
\usage{
clean(net, keep = c("id", "name", "gsymb"))
}
\arguments{
\item{net}{(Required) The igraph object that is supposed to loose all but the first two vertex attributes}

\item{keep}{(Optional) String-Vector that contains the vertex attributes you want to keep. Default is ("id", "name", "gsymb")}
}
\description{
Removes all vertex attributes except for default or user-specified attributes.
}
\examples{
annmods()
}
\keyword{helper-function}
