% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_and_farthest_modules.R
\name{closest_and_farthest_modules}
\alias{closest_and_farthest_modules}
\title{Module distance metrics}
\usage{
closest_and_farthest_modules(heatmap, g1_modules, g2_modules)
}
\arguments{
\item{heatmap}{(Optional) Array, containing the distances between modules}

\item{g1_modules}{(Optional) List of Strings, names of the modules from the first grapg}

\item{g2_modules}{(Optional) List of Strings, names of the modules from the second grapg}
}
\description{
Returns closest and farthest modules.
}
\examples{
annmods()
}
\keyword{helper-function}
\keyword{produces_graph,}
