% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTRL_example.R
\docType{data}
\name{CTRL_example}
\alias{CTRL_example}
\title{Example data of healthy volunteers}
\format{
A matrix with 31720 rows (genes) and 20 variables (healthy volunteers):
}
\source{
DEFINE SOURCE! (MAARS, pub, etc.)
}
\usage{
CTRL_example
}
\description{
Standardized microarray gene expression data of human skin from healthy volunteers.
}
\keyword{datasets}
