% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_overlap_size.R
\name{dist_overlap_size}
\alias{dist_overlap_size}
\title{Overlap between graphs}
\usage{
dist_overlap_size(net1, net2)
}
\arguments{
\item{net1}{(Required) igraph object, one of the two networks being compared}

\item{net2}{(Required) igraph object, the other of the two networks being compared}
}
\description{
Returns the size of the overlap (vertex count)
}
\examples{
annmods()
}
\keyword{helper-function}
