% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_jaccard_comp.R
\name{gene_jaccard_comp}
\alias{gene_jaccard_comp}
\title{gene_jaccard_comp
Compares two or multiple gene co-expression networks by the Jaccard index between their respective modules. Returns a distance matrix.}
\usage{
gene_jaccard_comp(in_graphs)
}
\arguments{
\item{in_graphs}{(Required) A named list of input graphs}
}
\description{
gene_jaccard_comp
Compares two or multiple gene co-expression networks by the Jaccard index between their respective modules. Returns a distance matrix.
}
\examples{
gene_mem_comp()
}
\keyword{module_comparison}
