% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mark_comp.R
\name{gene_mark_comp}
\alias{gene_mark_comp}
\title{gene_mark_comp}
\usage{
gene_mark_comp(in_graphs, markgens)
}
\arguments{
\item{in_graphs}{(Required) A named list of input graphs}

\item{markgens}{(Required) A vector of HGNC gene symbols, usually from the marker gene signature table used with CoMeAn.}
}
\description{
Compare marker gene percentages between input network modules
}
\examples{
gene_mark_comp()
}
\keyword{module_comparison}
