% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mem_comp.R
\name{gene_mem_comp}
\alias{gene_mem_comp}
\title{gene_mem_comp
Compares two or multiple gene co-expression networks by the number of shared genes between their respective modules.}
\usage{
gene_mem_comp(in_graphs)
}
\arguments{
\item{in_graphs}{(Required) A named list of input graphs}
}
\description{
gene_mem_comp
Compares two or multiple gene co-expression networks by the number of shared genes between their respective modules.
}
\examples{
gene_mem_comp()
}
\keyword{module_comparison}
