% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_perc_comp.R
\name{gene_perc_comp}
\alias{gene_perc_comp}
\title{gene_perc_comp
Compares two or multiple gene co-expression networks by the percentage of shared genes between their respective modules sum of genes between modules divided by the number of shared genes)}
\usage{
gene_perc_comp(in_graphs)
}
\arguments{
\item{in_graphs}{(Required) A named list of input graphs}
}
\description{
gene_perc_comp
Compares two or multiple gene co-expression networks by the percentage of shared genes between their respective modules sum of genes between modules divided by the number of shared genes)
}
\examples{
gene_perc_comp()
}
\keyword{module_comparison}
