% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_dist_e_weight.R
\name{hamming_dist_e_weight}
\alias{hamming_dist_e_weight}
\title{Returns the hamming-distance between two networks based off the sum of the edge-weights}
\usage{
hamming_dist_e_weight(net1, net2)
}
\arguments{
\item{net1}{(required): igraph object, one of the two networks being compared}

\item{net2}{(required): igraph object, the other of the two networks being compared}
}
\description{
Returns the hamming-distance between two networks based off the sum of the edge-weights
}
\examples{
annmods()
}
\keyword{helper-function}
