% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_dist_v_count.R
\name{hamming_dist_v_count}
\alias{hamming_dist_v_count}
\title{Returns the hamming-distance between two networks based off the number of shared vertices}
\usage{
hamming_dist_v_count(net1, net2)
}
\arguments{
\item{net1}{(required): igraph object, one of the two networks being compared}

\item{net2}{(required): igraph object, the other of the two networks being compared}
}
\description{
Returns the hamming-distance between two networks based off the number of shared vertices
}
\examples{
annmods()
}
\keyword{helper-function}
