% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_table.R
\name{information_table}
\alias{information_table}
\title{Produces a .gml and .csv file based off the passed along igraph object
Params:}
\usage{
information_table(graph, filename)
}
\arguments{
\item{net}{(Required) The igraph object is beign worked on}

\item{name}{(Required) Prefix for the produced table and .gml files}
}
\description{
Produces a .gml and .csv file based off the passed along igraph object
Params:
}
\examples{
annmods()
}
\keyword{produces_csv}
