% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_i.R
\name{jaccard_i}
\alias{jaccard_i}
\title{Jaccard similarity index between two lists}
\usage{
jaccard_i(a, b)
}
\arguments{
\item{a}{(Required) list of genenames from network or module A.}

\item{b}{(Required) list of genenames from network or module B.}
}
\description{
Given two character vectors, returns the Jaccard index.
}
\examples{
jaccard_i()
}
\keyword{co-expression}
\keyword{index}
\keyword{jaccard}
\keyword{network}
