% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_assignment.R
\name{module_assignment}
\alias{module_assignment}
\title{Module detection
Perform a clustering method on graph, returns a graph with module vertex attribute.}
\usage{
module_assignment(graph, method = "louvain")
}
\arguments{
\item{graph}{(Required) The igraph object is beign worked on}

\item{method}{(Optional) String, indicates the clustering method. Default: louvain}
}
\description{
Module detection
Perform a clustering method on graph, returns a graph with module vertex attribute.
}
\examples{
annmods()
}
\keyword{helper-function}
