% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_comparison.R
\name{module_comparision}
\alias{module_comparision}
\title{Calculates the hamming distance between two igraph objects}
\usage{
module_comparision(g1, g2, method = "Overlap")
}
\arguments{
\item{method}{(optional): String, sets the comparision algorithm. Default: "Overlap"}

\item{net1}{(required): One of the two igraph objects beign compared}

\item{net2}{(required): The other igraph object being compared}

\item{min_dist}{(required): float; value indicating if furher analysis should be done}
}
\description{
Calculates the hamming distance between two igraph objects
}
\examples{
annmods()
}
\keyword{helper-function}
