% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_comparison.R
\name{module_comparison}
\alias{module_comparison}
\title{Calculates the hamming distance between two igraph objects}
\usage{
module_comparison(g1, g2, dist_method = "overlap")
}
\arguments{
\item{dist_method}{(optional): String, sets the comparison algorithm. Default: "overlap"}

\item{net1}{(required): One of the two igraph objects beign compared}

\item{net2}{(required): The other igraph object being compared}
}
\description{
Calculates the hamming distance between two igraph objects
}
\examples{
annmods()
}
\keyword{helper-function}
