% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Returns the overlap between two networks, based off shared vertices}
\usage{
overlap(net1, net2)
}
\arguments{
\item{net1}{(required): igraph object, one of the two networks being compared}

\item{net2}{(required): igraph object, the other of the two networks being compared}
}
\description{
Returns the overlap between two networks, based off shared vertices
}
\examples{
annmods()
}
\keyword{helper-function}
