% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_overlap_graph.R
\name{produce_overlap_graph}
\alias{produce_overlap_graph}
\title{Reads in two files and produces a .gml file based off the overlap method}
\usage{
produce_overlap_graph(
  file1 = "AdL",
  file2 = "PsoL",
  cwd = "./PAP/data/",
  format1 = "gml",
  format2 = "gml"
)
}
\arguments{
\item{file1}{(optional): String, name of the file to be loaded. Default is "AdL"}

\item{file2}{(optional): String, name of the file to be loaded. Default is "PsoL"}

\item{cwd}{(optional): String, current working directory. Where to find the file. Default is "./PAP/data/"}

\item{format1}{(optional): String, specifies the file format. Default is "gml"}

\item{format2}{(optional): String, specifies the file format. Default is "gml"}
}
\description{
Reads in two files and produces a .gml file based off the overlap method
}
\examples{
ann_mods()
}
\keyword{helper-function}
\keyword{produces_graph,}
