% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSO_example.R
\docType{data}
\name{PSO_example}
\alias{PSO_example}
\title{Example data of psoriasis patients}
\format{
A matrix with 31688 rows (genes) and 20 variables (patients):
}
\source{
DEFINE SOURCE! (MAARS, pub, etc.)
}
\usage{
PSO_example
}
\description{
Standardized microarray gene expression data of human skin from psoriasis patients.
}
\keyword{datasets}
