% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{read_in}
\alias{read_in}
\title{Reads in the specified graph}
\usage{
read_in(name, cwd = "./", format = "gml", plot = FALSE)
}
\arguments{
\item{name}{(Required) String, the name of the file. It's important not to add the suffix!}

\item{cwd}{(Optional) String, current working directory. Where to find the file. Default is "./PAP/data/"}

\item{format}{(Optional) String, specifies the file format. Default is "gml"}

\item{plot}{(Optional) Boolean. Default is "FALSE". If TRUE, will plot the graph after loading.}
}
\description{
Reads in the specified graph
}
\examples{
annmods()
}
\keyword{helper-function}
