% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowMax.R
\name{RowMax}
\alias{RowMax}
\title{Receives a dataframe and returns the row max values as a vector}
\usage{
RowMax(data)
}
\arguments{
\item{data}{(required): Dataframe, contains the data to be used}
}
\description{
Receives a dataframe and returns the row max values as a vector
}
\examples{
ann_mods()
}
\keyword{helper-function}
