% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_plot.R
\name{save_and_plot}
\alias{save_and_plot}
\title{Helper function. Plots the passed .csv file and then saves the graph}
\usage{
save_and_plot(file1, file2, filename, hmap, cwd = "./PAP/data/out/")
}
\arguments{
\item{file1}{(Required) String, name of the first loaded file}

\item{file2}{(Required) String, name of the second loaded file}

\item{filename}{(Required) String, name of the file that the graph will be saved to}

\item{hmap}{(Required) The heatmap filename to output.}

\item{cwd}{(Optional) String, current working directory. Where to find the file. Default is "./PAP/data/out/"}
}
\description{
Helper function. Plots the passed .csv file and then saves the graph
}
\examples{
annmods()
}
\keyword{helper-function}
