% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_gene_analysis.R
\name{single_gene_analysis}
\alias{single_gene_analysis}
\title{Single gene and neigbourhood analysis}
\usage{
single_gene_analysis(graph, gene_name)
}
\arguments{
\item{graph}{(Required) The igraph object that's being analyized}

\item{gene_name}{(Required) Strig name of the vertex that will be analyzed}
}
\description{
Extracts a number of values for a specific vertex (read, gene) from the graph, returns a vector containing the neighborhood graph, as well as the gene's module membership, node degree, and hubscore.
Params:
}
\examples{
annmods()
}
\keyword{analysis}
\keyword{helper-function}
