% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkinSig_annotation.R
\docType{data}
\name{SkinSig_annotation}
\alias{SkinSig_annotation}
\title{Example data of gene annotations}
\format{
A data frame with 886 rows and 3 variables:
\describe{
  \item{HGNC.symbol}{Gene HGNC symbol}
  \item{Gene.description}{Functional annotation of gene}
  \item{SkinSig.signature}{General role and/or localization of variable expression}
}
}
\source{
DEFINE SOURCE! (MAARS, pub, etc.)
}
\usage{
SkinSig_annotation
}
\description{
A subset of genes differentially expressed between healthy, atopic dermatitis,
and psoriasis patients, and brief annotations of these genes.
}
\keyword{datasets}
