% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_gene_exprs.R
\name{target_gene_exprs}
\alias{target_gene_exprs}
\title{Target Gene Expression (CoMeAn)}
\usage{
target_gene_exprs(inlist, target_gene)
}
\arguments{
\item{inlist}{(Required) list of expression matrices}

\item{target_gene}{(Required) HGNC gene symbol (input matrix rowname) of the gene of interest}
}
\description{
Given a list of gene expression matrices, displays boxplot of gene expression per matrix.
}
\examples{
target_gene_exprs()
}
\keyword{co-expression}
\keyword{network}
