% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparision_intersection_only.R
\name{comparison_intersection_only}
\alias{comparison_intersection_only}
\title{Graph intersection comparison
Compares two graphs based on their overlap, based off datastructures found via unsupervised learning.}
\usage{
comparison_intersection_only(
  file1 = "AdL",
  file2 = "PsoL",
  min_dist = 0.75,
  cwd = "./PAP/data/",
  format1 = "gml",
  format2 = "gml",
  overlap = "Overlap"
)
}
\arguments{
\item{file1}{(Required) String, name of the file to be loaded. Default is "AdL"}

\item{file2}{(Required) String, name of the file to be loaded. Default is "PsoL"}

\item{min_dist}{(Optional) Float, value the edges/connctions need to larger than to be kept}

\item{cwd}{(Optional) String, current working directory. Where to find the file. Default is "./PAP/data/"}

\item{format1}{(Optional) String, specifies the file format. Default is "gml"}

\item{format2}{(Optional) String, specifies the file format. Default is "gml"}

\item{overlap}{(Optional) String, specifies the method used to calculate module distance. Default is "Overlap"}
}
\description{
Graph intersection comparison
Compares two graphs based on their overlap, based off datastructures found via unsupervised learning.
}
\examples{
annmods()
}
\keyword{analysis_function}
