% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Fisher_comp.R
\name{mod_Fisher_comp}
\alias{mod_Fisher_comp}
\title{mod_Fisher_comp
Fisher test on the cell/tissue type-annotated modules of two or multiple gene co-expression networks. Returns a data frame of significant enrichment per module.}
\usage{
mod_Fisher_comp(in_graphs, markgens)
}
\arguments{
\item{in_graphs}{(Required) A named list of input graphs}

\item{markgens}{(Required) A vector of HGNC gene symbols, usually from the marker gene signature table used with CoMeAn.}
}
\description{
mod_Fisher_comp
Fisher test on the cell/tissue type-annotated modules of two or multiple gene co-expression networks. Returns a data frame of significant enrichment per module.
}
\examples{
mod_Fisher_comp()
}
\keyword{module_comparison}
