% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{net_comparison}
\alias{net_comparison}
\title{net_comparison}
\usage{
net_comparison(file1 = "AdL", file2 = "PsoL", min_dist = 0.75,
  cwd = "./PAP/data/", format1 = "gml", format2 = "gml")
}
\arguments{
\item{file1}{(Required) String, name of the file to be loaded. Default is "AdL"}

\item{file2}{(Required) String, name of the file to be loaded. Default is "PsoL"}

\item{min_dist}{(Optional) Float, value the edges/connctions need to larger than to be kept}

\item{cwd}{(Optional) String, current working directory. Where to find the file. Default is "./PAP/data/"}

\item{format1}{(Optional) String, specifies the file format. Default is "gml"}

\item{format2}{(Optional) String, specifies the file format. Default is "gml"}
}
\description{
Compares two graphs based off datastructures found via unsupervised learning
}
\examples{
annmods()
TODO: add use for mode/add different modes
TODO: Complete Annotations (add return params)
TODO: Add method to analyze single gene
TODO: Way to combine V. count and node degree
TODO: More than two graphs
TODO: Extract methods
TODO: Change directory to where files get saved
Title     : net_comparison
Objective : Provide a number of functions to compare two (or more) networks
Created by: Nicholas Schmitt
}
\keyword{produces_plot}
