% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMax.R
\name{rowMax}
\alias{rowMax}
\title{Receives a dataframe and returns the row max values as a vector}
\usage{
rowMax(data)
}
\arguments{
\item{data}{(required): Dataframe, contains the data to be used}
}
\description{
Receives a dataframe and returns the row max values as a vector
}
\examples{
annmods()
}
\keyword{helper-function}
