% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowVar.R
\name{RowVar}
\alias{RowVar}
\title{Row-wise variance values of a matrix}
\usage{
RowVar(x, ...)
}
\arguments{
\item{x}{(Required) Input gene expression matrix.}
}
\description{
Row-wise variance values of a matrix
}
\examples{
RowVar()
}
\keyword{Biobase}
\keyword{clusterprofiler}
\keyword{co-expression}
\keyword{igraph}
\keyword{network}
\keyword{org.Hs.eg.db}
