% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_comparison.R
\name{self_comparison}
\alias{self_comparison}
\title{Generates new modules for an igraph object and then compares novel modules to the original modules}
\usage{
self_comparison(file, cwd = "./", format = "gml")
}
\arguments{
\item{file}{(Optional) String, name of the file to be loaded. Default is "AdL"}

\item{cwd}{(Optional) String, current working directory. Where to find the file. Default is "./PAP/data/"}

\item{format}{(Optional)  String, specifies the file format. Default is "gml"}
}
\description{
Generates new modules for an igraph object and then compares novel modules to the original modules
}
\examples{
ann_mods()
}
\keyword{produces_plot}
